+++ client/index.js
@@ -7,7 +7,7 @@ import { render } from '@wordpress/element';
  * Internal dependencies
  */
 import './stylesheets/index.scss';
-import Navigation from './navigation';
+import Navigation from './_navigation';
 
 const navigationRoot = document.getElementById(
 	'woocommerce-embedded-navigation'
